how-to-probe/pyproject.toml
[project]
name = "how-to-probe"
version = "0.1.0"
description = "How to probe install"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "captum==0.6.0",
    "einops==0.6.1",
    "imageio==2.33.1",
    "importlib-metadata==6.8.0",
    "lime==0.2.0.1",
    "matplotlib==3.7.2",
    "mmcv==2.0.1",
    "mmengine==0.8.4",
    "mmpretrain",
    "numpy==1.25.2",
    "opencv-python==4.8.0.74",
    "openmim==0.3.9",
    "openxlab==0.0.17",
    "ordered-set==4.1.0",
    "pandas==2.0.3",
    "pillow==10.0.0",
    "pycocotools==2.0.6",
    "scikit-image==0.22.0",
    "scikit-learn==1.3.2",
    "scipy==1.11.3",
    "seaborn==0.12.2",
    "submitit==1.4.5",
    "sympy==1.12",
    "timm==0.9.7",
    "torch==2.0.1",
    "torchaudio==2.0.2",
    "torchvision==0.15.2",
    "zennit==0.5.1",
]

[tool.uv.sources]
mmpretrain = { git = "https://github.com/open-mmlab/mmpretrain.git", rev = "827a216155d86afa54a9feb8cd5e078c18adb75b" }
